+++ packages/elements/share/src/__tests__/unit/components/analyticsSpec.tsx
@@ -1,9 +1,9 @@
 import {
-  buttonClicked,
+  shareTriggerButtonClicked,
   cancelShare,
-  copyShareLink,
+  copyLinkButtonClicked,
   screenEvent,
-  submitShare,
+  formShareSubmitted,
 } from '../../../components/analytics';
 import {
   ConfigResponse,
@@ -22,9 +22,9 @@ describe('share analytics', () => {
     })),
   });
 
-  describe('buttonClicked', () => {
+  describe('shareTriggerButtonClicked', () => {
     it('should create event payload', () => {
-      expect(buttonClicked()).toMatchObject({
+      expect(shareTriggerButtonClicked()).toMatchObject({
         eventType: 'ui',
         action: 'clicked',
         actionSubject: 'button',
@@ -66,9 +66,9 @@ describe('share analytics', () => {
     });
   });
 
-  describe('copyShareLink', () => {
+  describe('copyLinkButtonClicked', () => {
     it('should create event payload without origin id', () => {
-      expect(copyShareLink(100)).toMatchObject({
+      expect(copyLinkButtonClicked(100)).toMatchObject({
         eventType: 'ui',
         action: 'clicked',
         actionSubject: 'button',
@@ -83,7 +83,7 @@ describe('share analytics', () => {
 
     it('should create event payload with origin id', () => {
       const shareOrigin: OriginTracing = mockShareOrigin();
-      expect(copyShareLink(100, shareOrigin)).toMatchObject({
+      expect(copyLinkButtonClicked(100, shareOrigin)).toMatchObject({
         eventType: 'ui',
         action: 'clicked',
         actionSubject: 'button',
@@ -103,7 +103,7 @@ describe('share analytics', () => {
     });
   });
 
-  describe('submitShare', () => {
+  describe('formShareSubmitted', () => {
     const data: DialogContentState = {
       users: [
         {
@@ -128,7 +128,7 @@ describe('share analytics', () => {
       },
     };
     it('should create event payload without share content type and origin id', () => {
-      expect(submitShare(100, data)).toMatchObject({
+      expect(formShareSubmitted(100, data)).toMatchObject({
         eventType: 'ui',
         action: 'clicked',
         actionSubject: 'button',
@@ -149,7 +149,7 @@ describe('share analytics', () => {
     });
 
     it('should create event payload without origin id', () => {
-      expect(submitShare(100, data, 'issue')).toMatchObject({
+      expect(formShareSubmitted(100, data, 'issue')).toMatchObject({
         eventType: 'ui',
         action: 'clicked',
         actionSubject: 'button',
@@ -172,27 +172,29 @@ describe('share analytics', () => {
 
     it('should create event payload with origin id', () => {
       const shareOrigin: OriginTracing = mockShareOrigin();
-      expect(submitShare(100, data, 'issue', shareOrigin)).toMatchObject({
-        eventType: 'ui',
-        action: 'clicked',
-        actionSubject: 'button',
-        actionSubjectId: 'submitShare',
-        attributes: expect.objectContaining({
-          contentType: 'issue',
-          duration: expect.any(Number),
-          teamCount: 1,
-          userCount: 1,
-          emailCount: 1,
-          users: ['abc-123'],
-          teams: ['123-abc'],
-          packageVersion: expect.any(String),
-          packageName: '@atlaskit/share',
-          isMessageEnabled: false,
-          messageLength: 0,
-          originIdGenerated: 'abc-123',
-          originProduct: 'jest',
-        }),
-      });
+      expect(formShareSubmitted(100, data, 'issue', shareOrigin)).toMatchObject(
+        {
+          eventType: 'ui',
+          action: 'clicked',
+          actionSubject: 'button',
+          actionSubjectId: 'submitShare',
+          attributes: expect.objectContaining({
+            contentType: 'issue',
+            duration: expect.any(Number),
+            teamCount: 1,
+            userCount: 1,
+            emailCount: 1,
+            users: ['abc-123'],
+            teams: ['123-abc'],
+            packageVersion: expect.any(String),
+            packageName: '@atlaskit/share',
+            isMessageEnabled: false,
+            messageLength: 0,
+            originIdGenerated: 'abc-123',
+            originProduct: 'jest',
+          }),
+        },
+      );
       expect(shareOrigin.toAnalyticsAttributes).toHaveBeenCalledTimes(1);
       expect(shareOrigin.toAnalyticsAttributes).toHaveBeenCalledWith({
         hasGeneratedId: true,
@@ -206,7 +208,7 @@ describe('share analytics', () => {
         allowComment: true,
       };
       expect(
-        submitShare(100, data, 'issue', shareOrigin, config),
+        formShareSubmitted(100, data, 'issue', shareOrigin, config),
       ).toMatchObject({
         eventType: 'ui',
         action: 'clicked',
@@ -264,7 +266,7 @@ describe('share analytics', () => {
         allowComment: true,
       };
       expect(
-        submitShare(100, dataWithMembers, 'issue', shareOrigin, config),
+        formShareSubmitted(100, dataWithMembers, 'issue', shareOrigin, config),
       ).toMatchObject({
         eventType: 'ui',
         action: 'clicked',
@@ -305,7 +307,7 @@ describe('share analytics', () => {
       };
 
       expect(
-        submitShare(100, dataWithMembers, 'issue', shareOrigin, config),
+        formShareSubmitted(100, dataWithMembers, 'issue', shareOrigin, config),
       ).toMatchObject({
         eventType: 'ui',
         action: 'clicked',