+++ packages/elements/share/src/components/analytics.ts
@@ -45,7 +45,10 @@ const createScreenEvent = (
   attributes: buildAttributes(attributes),
 });
 
-export const buttonClicked = () =>
+export const screenEvent = () => createScreenEvent('shareModal');
+
+// = share dialog invoked. Not to be confused with "share submitted"
+export const shareTriggerButtonClicked = () =>
   createEvent('ui', 'clicked', 'button', 'share');
 
 export const cancelShare = (start: number) =>
@@ -54,19 +57,17 @@ export const cancelShare = (start: number) =>
     duration: duration(start),
   });
 
-export const copyShareLink = (
+export const copyLinkButtonClicked = (
   start: number,
   shareOrigin?: OriginTracing | null,
 ) =>
   createEvent('ui', 'clicked', 'button', 'copyShareLink', {
     source: 'shareModal',
     duration: duration(start),
-    ...handleShareOrigin(shareOrigin),
+    ...getOriginTracingAttributes(shareOrigin),
   });
 
-export const screenEvent = () => createScreenEvent('shareModal');
-
-export const submitShare = (
+export const formShareSubmitted = (
   start: number,
   data: DialogContentState,
   shareContentType?: string,
@@ -78,7 +79,7 @@ export const submitShare = (
   const teamUserCounts = extractMemberCountsFromTeams(data, isTeam);
   const emails = extractIdsByType(data, isEmail);
   return createEvent('ui', 'clicked', 'button', 'submitShare', {
-    ...handleShareOrigin(shareOrigin),
+    ...getOriginTracingAttributes(shareOrigin),
     contentType: shareContentType,
     duration: duration(start),
     emailCount: emails.length,
@@ -100,10 +101,8 @@ export const submitShare = (
 
 const duration = (start: number) => Date.now() - start;
 
-const handleShareOrigin = (shareOrigin?: OriginTracing | null) =>
-  shareOrigin
-    ? shareOrigin.toAnalyticsAttributes({ hasGeneratedId: true })
-    : {};
+const getOriginTracingAttributes = (origin?: OriginTracing | null) =>
+  origin ? origin.toAnalyticsAttributes({ hasGeneratedId: true }) : {};
 
 const extractIdsByType = <T extends OptionData>(
   data: DialogContentState,