+++ packages/navigation/atlassian-switcher/examples/81-uc-confluence-switcher-example.tsx
@@ -1,11 +1,11 @@
-import React, { Component } from 'react';
+import React from 'react';
 import Button from '@atlaskit/button';
 import Drawer from '@atlaskit/drawer';
-import { mockEndpoints, REQUEST_FAST } from './helpers/mock-endpoints';
+import { mockEndpoints, REQUEST_MEDIUM } from './helpers/mock-endpoints';
 import { withAnalyticsLogger, withIntlProvider } from './helpers';
 import AtlassianSwitcher from '../src';
 
-class JiraSwitcherExample extends Component {
+class ConfluenceSwitcherExample extends React.Component {
   state = {
     isDrawerOpen: false,
   };
@@ -16,25 +16,16 @@ class JiraSwitcherExample extends Component {
 
   openDrawer = () => {
     mockEndpoints(
-      'jira',
+      'confluence',
       originalMockData => {
         return {
           ...originalMockData,
           LICENSE_INFORMATION_DATA: {
-            hostname: 'https://some-random-instance.jira-dev.com',
-            firstActivationDate: 1492488658539,
-            maintenanceEndDate: '2017-04-24',
-            maintenanceStartDate: '2017-04-17',
-            products: {
-              'jira-software.ondemand': {
-                billingPeriod: 'ANNUAL',
-                state: 'ACTIVE',
-              },
-            },
+            notUsedInMode: 'user-centric',
           },
         };
       },
-      REQUEST_FAST,
+      REQUEST_MEDIUM,
     );
     this.setState({
       isDrawerOpen: true,
@@ -58,10 +49,10 @@ class JiraSwitcherExample extends Component {
       <div style={{ padding: '2rem' }}>
         <Drawer onClose={this.onClose} isOpen={this.state.isDrawerOpen}>
           <AtlassianSwitcher
-            product="jira"
+            product="confluence"
             cloudId="some-cloud-id"
+            enableUserCentricProducts
             triggerXFlow={this.onTriggerXFlow}
-            enableExpandLink={true}
           />
         </Drawer>
         <Button type="button" onClick={this.openDrawer}>
@@ -72,4 +63,4 @@ class JiraSwitcherExample extends Component {
   }
 }
 
-export default withIntlProvider(withAnalyticsLogger(JiraSwitcherExample));
+export default withIntlProvider(withAnalyticsLogger(ConfluenceSwitcherExample));