+++ packages/navigation/atlassian-switcher/src/components/generic-switcher.tsx
@@ -4,6 +4,7 @@ import Switcher from './switcher';
 import CommonDataProvider from '../providers/common-data-provider';
 import { Product, FeatureFlagProps } from '../types';
 import { mapResultsToSwitcherProps } from '../utils/map-results-to-switcher-props';
+import { AvailableProductsProvider } from '../providers/products-data-provider';
 
 type GenericSwitcherProps = {
   cloudId: string;
@@ -32,18 +33,28 @@ const getFeatures = (
 };
 
 export default (props: GenericSwitcherProps) => (
-  <CommonDataProvider cloudId={props.cloudId}>
-    {providerResults => {
-      const switcherLinks = mapResultsToSwitcherProps(
-        props.cloudId,
-        providerResults,
-        {
-          ...props.features,
-          ...getFeatures(props.product),
-        },
-      );
+  <AvailableProductsProvider
+    isUserCentric={props.features.enableUserCentricProducts}
+  >
+    {availableProducts => (
+      <CommonDataProvider
+        cloudId={props.cloudId}
+        isUserCentric={Boolean(props.features.enableUserCentricProducts)}
+      >
+        {providerResults => {
+          const switcherLinks = mapResultsToSwitcherProps(
+            props.cloudId,
+            providerResults,
+            {
+              ...props.features,
+              ...getFeatures(props.product),
+            },
+            availableProducts,
+          );
 
-      return <Switcher {...props} {...switcherLinks} />;
-    }}
-  </CommonDataProvider>
+          return <Switcher {...props} {...switcherLinks} />;
+        }}
+      </CommonDataProvider>
+    )}
+  </AvailableProductsProvider>
 );