+++ packages/navigation/atlassian-switcher/src/components/switcher.tsx
@@ -82,11 +82,6 @@ export default class Switcher extends React.Component<SwitcherProps> {
     triggerXFlow(key, 'atlassian-switcher', event, analyticsEvent);
   };
 
-  getExpandHref = (hostname: string) => {
-    const isStagingInstance = hostname.indexOf('.jira-dev.com') !== -1;
-    return `//start.${isStagingInstance ? 'stg.' : ''}atlassian.com`;
-  };
-
   render() {
     const {
       messages,
@@ -151,6 +146,7 @@ export default class Switcher extends React.Component<SwitcherProps> {
               >
                 <SwitcherItem
                   icon={<item.Icon theme="product" />}
+                  description={item.description}
                   href={item.href}
                 >
                   {item.label}