+++ packages/search/global-search/src/__tests__/unit/GlobalQuickSearch.test.tsx
@@ -55,7 +55,21 @@ describe('GlobalQuickSearch', () => {
       .prop('onSearchInput');
     onSearchInput({ target: { value: 'foo' } });
 
-    expect(searchMock).toHaveBeenCalledWith('foo', 0);
+    expect(searchMock).toHaveBeenCalledWith('foo', 0, undefined);
+  });
+
+  it('should handle searching with filter applied', () => {
+    const searchMock = jest.fn();
+    const filters = [{ '@type': 'spaces', spaces: ['TEST'] }];
+    const wrapper = render({ onSearch: searchMock, filters });
+
+    const onSearchInput: Function = wrapper
+      .children()
+      .first()
+      .prop('onSearchInput');
+    onSearchInput({ target: { value: 'foo' } });
+
+    expect(searchMock).toHaveBeenCalledWith('foo', 0, filters);
   });
 
   it('should fire searches with the queryVersion parameter incrementing', () => {
@@ -68,10 +82,10 @@ describe('GlobalQuickSearch', () => {
       .prop('onSearchInput');
 
     onSearchInput({ target: { value: 'foo' } });
-    expect(searchMock).toHaveBeenNthCalledWith(1, 'foo', 0);
+    expect(searchMock).toHaveBeenNthCalledWith(1, 'foo', 0, undefined);
 
     onSearchInput({ target: { value: 'foo' } });
-    expect(searchMock).toHaveBeenNthCalledWith(2, 'foo', 1);
+    expect(searchMock).toHaveBeenNthCalledWith(2, 'foo', 1, undefined);
   });
 
   it('should trim the search input', () => {
@@ -84,7 +98,7 @@ describe('GlobalQuickSearch', () => {
       .prop('onSearchInput');
     onSearchInput({ target: { value: '  pattio   ' } });
 
-    expect(searchMock).toHaveBeenCalledWith('pattio', 0);
+    expect(searchMock).toHaveBeenCalledWith('pattio', 0, undefined);
   });
 
   describe('Search result events', () => {