+++ packages/search/global-search/src/components/common/QuickSearchContainer.tsx
@@ -8,12 +8,12 @@ import {
 import GlobalQuickSearch from '../GlobalQuickSearch';
 import performanceNow from '../../util/performance-now';
 import {
-  GenericResultMap,
   ResultsWithTiming,
   Result,
   ResultsGroup,
   ConfluenceResultsMap,
   Results,
+  JiraResultsMap,
 } from '../../model/Result';
 import {
   ShownAnalyticsAttributes,
@@ -35,6 +35,7 @@ import {
 } from '../../util/features';
 import { Scope, QuickSearchContext } from '../../api/types';
 import { CONF_OBJECTS_ITEMS_PER_PAGE } from '../../util/experiment-utils';
+import { Filter } from '../../api/CrossProductSearchClient';
 
 const resultMapToArray = (results: ResultsGroup[]): Result[][] =>
   results.map(result => result.items);
@@ -42,10 +43,19 @@ const resultMapToArray = (results: ResultsGroup[]): Result[][] =>
 export interface SearchResultProps<T> extends State<T> {
   retrySearch: () => void;
   searchMore: (scope: Scope) => void;
+  onFilterChanged(filter: Filter[]): void;
+}
+
+export interface FilterComponentProps<T> {
+  latestSearchQuery: string;
+  searchResults: T | null;
+  isLoading: boolean;
+  currentFilters: Filter[];
+  onFilterChanged(filter: Filter[]): void;
 }
 
 export interface PartiallyLoadedRecentItems<
-  T extends ConfluenceResultsMap | GenericResultMap
+  T extends ConfluenceResultsMap | JiraResultsMap
 > {
   // Represents recent items that should be present before any UI is shown
   eagerRecentItemsPromise: Promise<ResultsWithTiming<T>>;
@@ -53,17 +63,19 @@ export interface PartiallyLoadedRecentItems<
   lazyLoadedRecentItemsPromise: Promise<Partial<T>>;
 }
 
-export interface Props<T extends ConfluenceResultsMap | GenericResultMap> {
+export interface Props<T extends ConfluenceResultsMap | JiraResultsMap> {
   logger: Logger;
   linkComponent?: LinkComponent;
   product: QuickSearchContext;
   getSearchResultsComponent(state: SearchResultProps<T>): React.ReactNode;
+  getFilterComponent(props: FilterComponentProps<T>): React.ReactNode;
   getRecentItems(sessionId: string): PartiallyLoadedRecentItems<T>;
   getSearchResults(
     query: string,
     sessionId: string,
     startTime: number,
     queryVersion: number,
+    filters: Filter[],
   ): Promise<ResultsWithTiming<T>>;
   getAutocompleteSuggestions?(query: string): Promise<string[]>;
   referralContextIdentifiers?: ReferralContextIdentifiers;
@@ -115,6 +127,7 @@ export interface State<T> {
   searchResults: T | null;
   recentItems: T | null;
   autocompleteSuggestions?: string[];
+  currentFilters: Filter[];
 }
 
 const LOGGER_NAME = 'AK.GlobalSearch.QuickSearchContainer';
@@ -122,11 +135,15 @@ const LOGGER_NAME = 'AK.GlobalSearch.QuickSearchContainer';
  * Container/Stateful Component that handles the data fetching and state handling when the user interacts with Search.
  */
 export class QuickSearchContainer<
-  T extends ConfluenceResultsMap | GenericResultMap
+  T extends ConfluenceResultsMap | JiraResultsMap
 > extends React.Component<Props<T>, State<T>> {
   // used to terminate if component is unmounted while waiting for a promise
   unmounted: boolean = false;
   latestQueryVersion: number = 0;
+
+  static defaultProps = {
+    getFilterComponent: () => null,
+  };
 
   constructor(props: Props<T>) {
     super(props);
@@ -138,6 +155,7 @@ export class QuickSearchContainer<
       recentItems: null,
       searchResults: null,
       keepPreQueryState: true,
+      currentFilters: [],
     };
   }
 
@@ -171,7 +189,7 @@ export class QuickSearchContainer<
     this.unmounted = true;
   }
 
-  doSearch = async (query: string, queryVersion: number) => {
+  doSearch = async (query: string, queryVersion: number, filters: Filter[]) => {
     const startTime: number = performanceNow();
     this.latestQueryVersion = queryVersion;
 
@@ -181,6 +199,7 @@ export class QuickSearchContainer<
         this.state.searchSessionId,
         startTime,
         queryVersion,
+        filters,
       );
 
       if (this.unmounted) {
@@ -326,13 +345,21 @@ export class QuickSearchContainer<
     }
   };
 
-  handleSearch = (newLatestSearchQuery: string, queryVersion: number) => {
-    if (this.state.latestSearchQuery === newLatestSearchQuery) {
+  handleSearch = (
+    newLatestSearchQuery: string,
+    queryVersion: number,
+    filters: Filter[],
+  ) => {
+    if (
+      this.state.latestSearchQuery === newLatestSearchQuery &&
+      filters === this.state.currentFilters
+    ) {
       return;
     }
 
     this.setState({
       latestSearchQuery: newLatestSearchQuery,
+      currentFilters: filters,
       isLoading: true,
     });
 
@@ -343,18 +370,23 @@ export class QuickSearchContainer<
           isError: false,
           isLoading: false,
           keepPreQueryState: true,
+          currentFilters: [],
         },
         () => {
           this.fireShownPreQueryEvent();
         },
       );
     } else {
-      this.doSearch(newLatestSearchQuery, queryVersion);
+      this.doSearch(newLatestSearchQuery, queryVersion, filters);
     }
   };
 
   retrySearch = () => {
-    this.handleSearch(this.state.latestSearchQuery, this.latestQueryVersion);
+    this.handleSearch(
+      this.state.latestSearchQuery,
+      this.latestQueryVersion,
+      this.state.currentFilters,
+    );
   };
 
   async componentDidMount() {
@@ -474,9 +506,18 @@ export class QuickSearchContainer<
     }
   };
 
+  handleFilter = (filter: Filter[]) => {
+    this.handleSearch(
+      this.state.latestSearchQuery,
+      this.latestQueryVersion,
+      filter,
+    );
+  };
+
   render() {
     const {
       linkComponent,
+      getFilterComponent,
       getSearchResultsComponent,
       placeholder,
       selectedResultId,
@@ -492,6 +533,7 @@ export class QuickSearchContainer<
       recentItems,
       keepPreQueryState,
       autocompleteSuggestions,
+      currentFilters,
     } = this.state;
 
     return (
@@ -507,7 +549,15 @@ export class QuickSearchContainer<
         onSelectedResultIdChanged={onSelectedResultIdChanged}
         inputControls={inputControls}
         autocompleteSuggestions={autocompleteSuggestions}
+        filters={this.state.currentFilters}
       >
+        {getFilterComponent({
+          onFilterChanged: this.handleFilter,
+          searchResults,
+          currentFilters,
+          isLoading,
+          latestSearchQuery,
+        })}
         {getSearchResultsComponent({
           retrySearch: this.retrySearch,
           latestSearchQuery,
@@ -518,6 +568,8 @@ export class QuickSearchContainer<
           keepPreQueryState,
           searchSessionId,
           searchMore: this.getMoreSearchResults,
+          currentFilters,
+          onFilterChanged: this.handleFilter,
         })}
       </GlobalQuickSearch>
     );